Make the non-scrambled AX.25 decoder more conservative
In case of scrambling the self synchronizing scrambler ensures
that enough repetitions of the AX.25 flag have been received.
However, this does not hold for the case of non scrambled
transmissions. In this case, we wait for at least two consecutive
AX.25 flags to reduce the false alarms. Experiments have shown
that due to the poor CRC there were many false positive frames.

It is expected however, to miss some transmissions that use only one
AX.25 flag.

+++ include/satnogs/ax25_decoder.h
@@ -127,7 +127,7 @@ private:
   const bool d_crc_check;
   const size_t d_max_frame_len;
   decoding_state_t d_state;
-  uint8_t d_shift_reg;
+  uint16_t d_shift_reg;
   uint8_t d_dec_b;
   uint8_t d_prev_bit_nrzi;
   size_t d_received_bytes;
